Clean up the Functions api.
This introduces a formal `FunctionDefinition` that contains the implementation name of the function and a flag whether its an aggregate or not.

This definition is used for the build in scalars, aggregates and predicates now. It can be used in the future for user defined functions (udfs) as well. Furthermore it could cater for better static definitions, like argument cardinality and type.

It also adds a test for and closes #44.