Avoid star and static imports.
This changes enables checkstyle rules for avoid static but especially star imports.

While I do think that star imports doesn’t do much harm in applications, it has the potential to do big harm in a library such as this. For a backround read this: http://www.javadude.com/posts/20040522-import-on-demand-is-evil/.

Also, my sense of beautity to prefer `update(CREATE, pattern);` in contrast to `update(UpdateType.CREATE, pattern);` for example is much worth compared to actual hassle.