Automatically generate symbolic name if required.
This commit changes the named Node and Relationship in that regard that a request for the required symbolic name generates an „unresolved“ symbolic name. Unresolved in that context means that the generated object is just a placeholder with no value.

That value can be generated in place during the visitation of the AST. With the currenty implementation of the rendering vistor, there’s context for that.

This closes #70.