GH-93 - Remove usage of MethodHandles and support native compilation on GraalVM.
This commit removes the usage of `MethodHandle` in the `RenderingVisitor`. Instead it uses plain `Methods`. It also adds the necessary meta information to open up the visitors methods for reflection access on GraalVM.

Method handles are only supported in some cases on GraalVM and giving the observations done by Geoffrey https://www.optaplanner.org/blog/2018/01/09/JavaReflectionButMuchFaster.html, don't provide much benefit in our scenario as well.

A new project has been added that takes a simple application, compiles it into a binary and runs an integration test on that binary to see if produces some expected output.