GH-121 - Don't force rendering of aliases when the original object is used.
This changes prevents the forced rendering of an alias for objects when the original object - the one that has been aliased - is passed down to the DSL after an alias has been created.

The original intention for that behaviour was related to Map projection, in which the alias is actually rendered before the object. That behaviour can be assured in a better way.

So now the use of an aliased expression the first time triggers `a AS b` respectively `b: a` in a map projection.
All further calls will just render `b`.

If the original object is used again, `a` will be rendered. As this can be an issue for some users relying on the old behaviour, this must appear in the release notes and requires at least a patch to 2020.2 if not even 2021.0.

This fixes #121.