GH-131 - Add support for projections on symbolic names.
According to the official Cypher manual map projections requires a (symbolic) name in all cases:

> A map projection begins with the variable bound to the graph entity to be projected from, and contains a body of comma-separated map elements

From https://neo4j.com/docs/cypher-manual/current/syntax/maps/#cypher-map-projection

Therefor the logical provider of methods that create a map projection is in fact the symbolic name.
This commit moves the logic away from both `Node` and `Relationship` into the `SymbolicName` and delegate to it.

The test added in #132 works as expected and only needs to use the static `Cypher#properties` to extract values from the name, unrelated to the actual issue.

This closes #131.