GH-155 - Provide infrastructure for generating a static meta model.
This adds a new module, `neo4j-cypher-dsl-codegen-core` to the project providing a public API that collections information about Nodes (their labels and properties) and Relationships (their types and properties) and their inner relations.

Two interfaces - `NodeModelBuilder` and `RelationshipModelBuilder` are provided as entrypoints together with a couple of supporting data classes (for collection properties as well as relationships).

In addition a small configuration framework has been provided that allows to parameterize those model builders with a fluent API.

This modul needs higher level API, for example an annotation processor, that reads information and generates the required information.

This closes #155.