Add support for some Query-DSL expressions.
This commit introduces the concept of a `ForeignAdapter` that allows to convert foreign expressions into expressions usable by the Cypher-DSL.
Together with that interface one implementation is provided, the `QueryDSLAdapter`.

The `QueryDSLAdapter` comes with Cypher templates, a Query-DSL visitor and everything that is needed to take many Query-DSL predicates and turn them into conditions usable by the Cypher-DSL.

All previous examples of the Cypher-DSL 3.3 compile correctly with that adapter.

The adapter does not strive a 100% translation from one DSL to another but should provide a sensible subset of what is possible so that many expressions can be written down via a typesafe Query-DSL model.