Ensure immutability of `RelationshipChain`. [bug]
The `RelationshipChain` was not immutable unlike all other completly build pattern elements. While this can be a breaking change for some users that didn’t use the returned value in that particular case, it poses more danger for the rest of the users who assumed a fresh instance was returned and a statement - once build - wouldn’t change anymore.