GH-180 - Add support for LOAD CSV and friends.
For this change to work, the default statement builder has been reusued
and inherited from. The ongoing call has been removed and instead,
additional AST elements will be added to the statement builder, either
via the procedure call builder or the new builder for LOAD CSV related
constructs.

This closes #180.