GH-192 - Don't introduce new symbolic names in conditional pattern expressions.
This commit checks if the visitor is currently in a relationship pattern condition (a conditional pattern expression). If so, a flag is set and the deque of vistited named elements is consulted if the elements of the pattern have already been seen in the scope and their symbolic names can be renderer or not.

To make this work, the `RelationshipPatternCondition` has been made public, internal API and moved to the `internal` package (which is not accessible when running on the module path).

This fixes #192.