GH-197 - Fix eagerly resolved symbolic names in negated pattern conditions.
The problem was manigfold:

- The `ExcludedPattern` had not been taken care of when checking if a pattern has been used as a condition
- We must also check vor visited named things by using only the symbolic name, it might be that someone used another anyNode with the same name.
- The pretty printing visitor didn’t (and couldn’t) call `enter(Condition)`, thus not marking a node as visited.

The issue actually required the use of two different renderes to appear:
Each rendered statement is cached by config.
Thus, the first call had always been correct, while the symbolic names had been resolved (but not rendered in case of the excluded pattern).
In a 2nd call with a different renderer, the names would be rendered as they didn’t need to be resolved anymore.

This fixes #197.