Create a cypher-dsl-parser module.
This commit introduces the new `org.neo4j:neo4j-cypher-dsl-parser` module. It
is based on `org.neo4j:neo4j-cypher-javacc-parser` in version 4.3.2, used by
Neo4j (the database) itself. Thus, all syntactically valid Cypher as of 4.3.2
can be parsed.

The main entry point to this module, `org.neo4j.cypherdsl.parser.CypherParser`,
uses a new AST factory that creates elements used throughout the Cypher-DSL
itself.

It provides dedicated methods for parsing expressions, clauses or full
statements. Expressions are handy to enrich or augment the Cypher-DSL builder,
clauses can be linked together by the new factory method
`org.neo4j.cypherdsl.core.Statement#of` which creates a statement with the
clauses in order they where passed on the factory and last but not least, full
statements can be used to create new union queries or in a subquery.

As the Cypher-DSL AST elements are meant to be immutable, no API is provided to
change generated elements. However, several filter and callbacks can be passed
into the parser via an `Options` object. Those filter or callbacks will receive
the freshly created elements in a way that their information can be used to
create other instances, exhibiting the desired characteristics.

In regards of dependencies the decision has been made to shade the Neo4j parser
and its transitive dependencies. While this leads to a rather big artifact as
of now (due to the 4.3 parser being dependent on Scala), it avoids version
clashes of both Neo4j and Scala when used in scenarios such as embedded.

This closes #181