fix: Make generated static model usable with self referential associations. (#337)
Prior to this fix, self referential relationships would lead to a situation in which the instance fields of those relationships in the static owning side of the model would either lead to a `NullpointerException` or an `StackOverflowError` on class initialization.

The solution to this problem is to not have those fields but provide explicit withers for all self referential associations.  As an example: A person following other people will no longer spot

```
public final Follows_ FOLLOWS = new Follows_(this, Person_.PERSON);
```

but instead have

```
public Follows_ withFollows(Person_ follows) {
	return new Follows_(this, follows);
}
```

so that a relation must be retrieved in an explicit way. There are test demonstrating this (see `org.neo4j.cypherdsl.examples.sdn6.UsageTest`).

The reasoning behind making it explicit is simple: Having self referential fields as instance field defined will eventually lead to a huge graph of instances if we want to make sure those can be modified individually or renaming, changing etc. will come with interesting surprises.

This fixes #335.