fix: Introduce dedicated schema name support for sanitising and quoting labels and types.
This introduces `neo4j-cypher-dsl-schema-name-support`, a module containing exactly one class, dedicated to sanitising and quoting schema names that are meant to be used as types or labels. It is guaranteed to produce sanitized names that are safe to be use in all Neo4j versions and will provide consistent results in the latest Neo4j version by default and with some configuration, the same results in older versions.