feat: Introduce standalone schema-name support module. (#445)
This introduces `neo4j-cypher-dsl-schema-name-support`, a module containing exactly one class, dedicated to sanitising and quoting schema names that are meant to be used as types or labels. It is guaranteed to produce sanitised and potentially quoted names that are safe to be used in string concatenation for creating labels and types in all Neo4j versions. In addition they will provide consistent results when working with certain unicode constellations. The latter requires a bit of knowledge about the Neo4j version used. The latest versions as of writing (4.4) is assumed by default, older need some help in some cases.

Thanks @AzuObs @AlexNeo4J and @robsdedude for your feedback on this work.