feat: Add support for the `COUNT {}` sub-query expressions. (#546)
This adds support for create Neo4j 5.x compatible `COUNT {}` sub-query expressions to the DSL. Expressions of that type can be built via `Expressions#count`. This change adds the general infrastructure for this feature and supports all the use-cases from the official Neo4j documentation. Anything not working or missing will be added in subsequent changes.

Closes #525.