refactor: Replace `requires static transitive` with `requires static`.
While this leads to the compiler complaining about missing modules during exports, it seems the only working way:
In a consumer project, the `static` part is ignored with the above declaration and things fail to compile without
the driver and friends being on the classpath.

The Java spec says in 7.7:

```
RequiresModifier:
(one of)
transitive static
```

"One of" seems to indicate that only one is allowed, but the compiler happily takes both.

https://docs.oracle.com/javase/specs/jls/se17/html/jls-7.html#jls-7.7