fix: Move resolved symbolic names into `StatementContext`.
This is necessary to keep the generated values constant for the lifetime of a statement.
Otherwise they would be re-resolved by the rendering when the cache overflew. This is probably
not what happened in #585, but what I realised while adding a test for the desired behaviour.

Having an API to access the resolved values might be useful in the future, too.

Fixes #585.