feat: Provide a catalog for identifiable items in a statement. (#617)
This adds a "catalog" feature for the both built and parsed statements. The catalog can be used to retrieve all labels and types as well as all resolved properties and the conditions those properties have been used in.

The catalog also contains the identifiable expressions at the end of a statement, so the corresponding method on the statement itself has been deprecated in will be removed in the next major version.

The catalog is built via a new visitor, which is also based on the existing reflective visitor. For better alignment of the API and proper restrict internals, the statement context has been moved into the internal package.

The catalog itself might contain more information in the future.