feat: Add label existences conditions to catalog. (#622)
This changes implements #618. The Neo4j documentation speaks about filters and we think it’s a better name to be used in our catalog too. So to unify the access to both property based filters (which filter by conditions) and label existence based filters, the `Condition` type has been renamed `PropertyFilter` and an interface `Filter` has been introduced, together with all required accessors for the label filters.

This closes #618.