feat: Add extended meta data and the ability to use generated variables. (#631)
A new option has been introduced to the rendering configuration that uses generated names for all symbolic names introduced for nodes and relationships as well as for parameters.

This will ensure constant rendering of statements regardless their source and how they are constructed.

The catalog introduced with cac7aea has been extended to be able to return a map from original parameter names to the generated ones. In addition, the parameter information directly on the statement has been deprecated and been moved in the the catalog.

This is not the full semantic analysis requested in #596, but some tests have added to achieve what is requested in that issue and therefor this closes #596 for the time being.

As one building block of a full analysis, `TreeNode.from` has been added, that allows to build an externally navigable tree from a statement so that no internal visitors are necessary when analysing a statement. This tree could be the root of a new module building full analysis.