feat: Provide a single DSL API entry point. (#862)
This makes `Cypher` the main entry point into Cypher-DSL: `Cypher` does now delegate to the corresponding classes, listed in the original ticket such as 

* `Conditions` and `Predicates`
* `Expressions`
* `Functions`
* `Operations`

Mostly all methods are straight forward replacement, mutators for adding and removing labels clearly indicate `label` in their name and for the sub-query with `with` we needed to introduce an overload that doesn't have a conflicting overload.

The old classes are still public but are marked as deprecated. They will be made package private in the next major release.