feat: Introduce Dialect 5.26 (#1185)
When selecting this dialect, Cypher-DSL ensures that all statements are prefixed with `CYPHER 5` when rendered to ensure compatiblity with Neo4j 2025.x in the future to come without jumping on Cypher 25 right now.

This has not been implemented as a visitor handler, but as post rendering operations to cater for nested statements such as subqueries, decorated queries and use queries. It would be a lot more work to figure out if those has been already prefixed or not.

Closes #1153.