refactor: Remove deprecated features and unnecessary suppressions.
For the next immanent major release, we must remove all current deprecations.
The only one left is `id` and friends, as Neo4j the database is still undecided about that.
We also removed all unnecessary suppresions of deprecations: They had been forgotten on the 2024 major release, in which we moved from multiple namespaced entry points into the DSL into a single one.