feat: Add an annotation processor for Neo4j-OGM annotations. (#1287)
This adds a second annotation processor for the annotations present in Neo4j-OGM, which behaves as close as possible as the SDN6+ annotation processor and will derive the same meta model for mappings that are semantically the same.

To cater for both processors, part of the processing infrastructure has been moved to the shared core module.

Additionally we figured out that the original processor that has been written to support both Java 8 and 11 would fail on Java 17 records. This feature has been added as well.

As OGM is widely used in Quarkus application together with the appropriate extension provided by Neo4j, examples have been added to showcase the setup of a Quarkus + Processing module. The example makes use of the latest Neo4j image, demonstrating the usage of dynamic labels in Cypher queries, which go nice with the meta model.

Formatting will further tackled in #1293. This commit closes #1286 and the annotation processor keeps being labeled as experimental, so that we can incorporate adjustments depending on the users feedback.

---------

Signed-off-by: Michael Simons <michael.simons@neo4j.com>
Co-authored-by: cquadflieg <cquadflieg@buhck.de>
Co-authored-by: Michael Simons <michael.simons@neo4j.com>