refactor: Use Spring-Java-Formatter and corresponding checkstyle config.
This unifies the formatting of this project with the Neo4j-JDBC driver, Spring Data Neo4j and others by applying the Spring Java Formatter und adding the corresponding build steps. The latter will break on wrongly formatted code, and can be fixed with the built-in goal `spring-javaformat:apply`.

This closes #1293

Signed-off-by: Michael Simons <michael.simons@neo4j.com>