refactor: Remove the remaining bits of driver dependencies.
The support for executable statements has already been removed, this removes the remaing value adapter. The adapter has never been used from SDN point of view (in contrast to the Query-DSL adapter). Keeping it ties the Cypher-DSL still to a specific driver version. With the rewrite of the value system at some point in driver post 6, this will be increasingly hard for something that is rather easy done in downstream applications.

It also untangles SDN driver versions of Cypher-DSL versions.

Last but not least, dependency exclusions in JDBC driver will be simplified. The latter depends in the translator module on Cypher-DSL, but ofc does not want to do anything with the common Java driver.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>