feat: Add support for dynamic label expressions. (#1323)
This adds support for `$()` expressions in labels, taking in either Cypher literals or Cypher parameters.

As it turned out during development of this feature, it must be an extension of the `LabelExpression` that we already had. The design of the latter was not good: It is a record and exposes basically everything.

This type has now been deprecated in favour of `Labels` with accompanying factory methods. A handful of methods on `Cypher` that used to take an instance of this type has also been deprecated but will keep on working by adapting the old type to the new one.

Dynamic label expression can now be parsed as well.

Closes #1314

Signed-off-by: Michael Simons <michael.simons@neo4j.com>