feat(codegen): Add option `org.neo4j.cypherdsl.codegen.excludes` for excluding types from processing.
This allows all of `@Node` and `@RelationshipProperties` (SDN)  as well as `@NodeEntity` and `@RelationshipEntity` (OGM) types from being excluded from annotation processing. A config option was used to avoid introducing another annotation that must work for both projects (SDN and OGM). Such an annotation would at least cause nasty dependency issues when not being shipped with an additional Cypher-DSL module, and we felt that would be a bit too much for that feature.

If a node annotated class is excluded then
- no model for that node will be generated
- no fields in other node models for relationships pointing to the excluded node will be generated
- if the node was part of a relationship property, that property won’t be generated either

If a relationship property class is excluded then
- no model for that entity will be generated
- no fields in other node models will be generated

This implements and closes #1350.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>